requirements.txt
annotated-types==0.7.0
APScheduler==3.10.4
arrow==1.4.0
attrs==25.4.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.0.0
google-auth==2.49.1
google-genai==0.3.0
gunicorn==25.1.0
ics==0.7.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==11.3.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0b2
pydantic_core==2.42.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.31.0
six==1.17.0
TatSu==5.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
websockets==14.2
Werkzeug==3.1.6
psycopg2-binary==2.9.11